[v4] Draft completed
