+++ src/action.ts
@@ -1,8 +1,19 @@
-import { dispatch } from "./subscriber";
+import { receive, send, blank } from "./chan";
+import { prop } from "./prop";
 
-export const action = () => {
-  function handler(...args: any[]) {
+const ActionChan = Symbol("ActionChan");
+const propActionChan = prop(ActionChan, blank);
+
+export function action() {
+  return function handler(...args: any[]) {
     dispatch(handler, ...args);
   }
-  return handler;
+}
+
+export function dispatch(target: any, ...args: any[]) {
+  send(propActionChan(target), args);
+}
+
+export function on(target: any, callback: (...args: any[]) => void) {
+  receive(propActionChan(target), (args: any[]) => callback(...args));
 }
