+++ src/index.ts
@@ -1,11 +1,10 @@
-export { useProvide } from "./useProvide";
-export { useSubscribe } from "./useSubscribe";
-export { action } from "./action";
-export { dispatch, subscribe } from "./subscriber";
-export { create, get, set } from "./store";
-export { provide, assign, override, cleanup } from "./di";
-export { modify } from "./immutable";
-export { reset } from "./reset";
+export { multi, chan, receive, send, group } from "./chan";
+export { store, get, set, modify, extend, watch, update } from "./store";
+export { action, dispatch, on } from "./action";
+export { service, assign, override, cleanup } from "./di";
 export { ssr, serialize, unserialize, register } from "./ssr";
+export { reset } from "./reset";
+export { useService } from "./useService";
+export { useSubscribe } from "./useSubscribe";
 export { useUnserialize } from "./useUnserialize";
 export { Subscribe } from "./Subscribe";
