+++ src/prop.ts
@@ -5,6 +5,9 @@ export function prop(prop: PropertyKey, _default?: () => any) {
     if (args.length === 2) {
       return target[prop] = _value;
     }
-    return target[prop] = target[prop] || (_default && _default());
+    if (!target.hasOwnProperty(prop)) {
+      target[prop] = _default && _default();
+    }
+    return target[prop];
   }) as ((target: any, _value?: any) => any);
 }
