+++ src/reset.ts
@@ -1,7 +1,7 @@
 import { reset as resetDi } from "./di";
 import { reset as resetSsr } from "./ssr";
 
-export const reset = () => {
+export function reset() {
   resetDi();
   resetSsr();
-};
+}
