+++ src/useUnserialize.ts
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 import { unserialize } from "./ssr";
 
-export const useUnserialize = (data: any) => {
+export function useUnserialize(data: any) {
   useMemo(() => {
     unserialize(data);
   }, [data]);
