--- src/v4.ts
@@ -1,3 +1,0 @@
-export { multi, chan, receive, send, group } from "./v4/chan";
-export { store, get, set, modify, extend, watch, update } from "./v4/store";
-export { action, dispatch, on } from "./v4/action";
