--- src/v4/reset.ts
@@ -1,7 +1,0 @@
-import { reset as resetDi } from "./di";
-import { reset as resetSsr } from "./ssr";
-
-export const reset = () => {
-  resetDi();
-  resetSsr();
-};
