+++ tests/action.test.ts
@@ -1,9 +1,17 @@
-import { action, subscribe } from "../src";
+import { action, on, dispatch } from "../src";
 
-test("Should work action", () => {
+test("Should work action as function", () => {
   const spy = jest.fn();
   const a = action();
-  subscribe(a, spy);
+  on(a, spy);
   a("A", 2);
+  expect(spy).toBeCalledWith("A", 2);
+});
+
+test("Should work action disptach", () => {
+  const spy = jest.fn();
+  const a = action();
+  on(a, spy);
+  dispatch(a, "A", 2);
   expect(spy).toBeCalledWith("A", 2);
 });
