+++ tests/useService.test.tsx
@@ -1,50 +1,50 @@
 import React from "react";
 import { mount } from "enzyme";
 import { act } from "react-dom/test-utils";
-import { useProvide, provide } from "../src";
+import { useService, service, set } from "../src";
 
 test("Should update component with useProvide", () => {
   class A {
-    store = "D";
+    state = "D";
   }
   const C = () => {
-    const a = useProvide(A);
-    return <p>{a.store}</p>
+    const a = useService(A);
+    return <p>{a.state}</p>
   };
 
   const el = mount(<C/>);
   expect(el.find("p").text()).toBe("D");
   act(() => {
-    provide(A).store = "DD";
+    set(service(A), "DD");
   });
   expect(el.find("p").text()).toBe("DD");
 });
 
 test("Should update component with array in useProvide", () => {
   class A {
-    store = "D";
+    state = "D";
   }
   class B {
-    store = "E";
+    state = "E";
   }
   const C = () => {
-    const [a, b] = useProvide([A, B]);
-    return <p>{a.store}{b.store}</p>
+    const [a, b] = useService([A, B]);
+    return <p>{a.state}{b.state}</p>
   };
 
   const el = mount(<C/>);
   expect(el.find("p").text()).toBe("DE");
   act(() => {
-    provide(A).store = "DD";
+    set(service(A), "DD");
   });
   expect(el.find("p").text()).toBe("DDE");
   act(() => {
-    provide(B).store = "EE";
+    set(service(B), "EE");
   });
   expect(el.find("p").text()).toBe("DDEE");
   act(() => {
-    provide(A).store = "A";
-    provide(B).store = "B";
+    set(service(A), "A");
+    set(service(B), "B");
   });
   expect(el.find("p").text()).toBe("AB");
 });
