+++ tests/useSubscribe.test.tsx
@@ -1,32 +1,15 @@
 import React, { useMemo } from "react";
 import { mount } from "enzyme";
-import { act } from "react-dom/test-utils";
-import { provide, create, useProvide, useSubscribe } from "../src";
-
-test("Should update component with useProvide", () => {
-  class A {
-    store = "D";
-  }
-  const C = () => {
-    const a = useProvide(A);
-    return <p>{a.store}</p>
-  };
-  const el = mount(<C/>);
-  expect(el.find("p").text()).toBe("D");
-  act(() => {
-    provide(A).store = "DD";
-  })
-  expect(el.find("p").text()).toBe("DD");
-});
+import { store, useSubscribe, set } from "../src";
 
 test("Should update component with useSubscribe", () => {
   class A {
-    store = "D";
+    state = "D";
   }
   const C = () => {
-    const a = useMemo(() => create(A), []);
+    const a = useMemo(() => store(A), []);
     useSubscribe(a);
-    return <p onClick={() => a.store = "DD"}>{a.store}</p>
+    return <p onClick={() => set(a, "DD")}>{a.state}</p>
   };
   const el = mount(<C/>);
   expect(el.find("p").text()).toBe("D");
